fix: slt/i/u instruction implemention

+++ with_emulator/src/emulator.h
@@ -509,7 +509,7 @@ public:
       } else {
         results[0] = (instruction >> 30) ? (rs1 - rs2) : (rs1 + rs2);
         results[1] = (rs1 << (rs2 & 63));
-        results[2] = (( (signed) rs1) < ( (signed) rs2));
+        results[2] = (( (signed long) rs1) < ( (signed long) rs2));
         results[3] = (rs1 < rs2);
         results[4] = (rs1 ^ rs2);
         results[5] = (instruction >> 30) ? (((signed long) rs1) >> (rs2 & 63)) : (rs1 >> (rs2 & 63));
@@ -526,7 +526,7 @@ public:
       #endif
       results[0] = (rs1 + immediate_i_type);
       results[1] = (rs1 << (immediate_i_type & 63));
-      results[2] = (( (signed) rs1) < ( (signed) immediate_i_type));
+      results[2] = (( (signed long) rs1) < ( (signed long) immediate_i_type));
       results[3] = (rs1 < immediate_i_type);
       results[4] = (rs1 ^ immediate_i_type);
       results[5] = (instruction & (1 << 30)) ? (((signed long) rs1) >> (immediate_i_type & 63)) : (rs1 >> (immediate_i_type & 63));
